move snapshots
