+++ pkg/module_manager/go_hook/go_hook.go
@@ -1,12 +1,6 @@
 package go_hook
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"reflect"
-	"strings"
 	"time"
 
 	"github.com/deckhouse/deckhouse/pkg/log"
@@ -42,64 +36,6 @@ type HookMetadata struct {
 	Global         bool
 	EmbeddedModule bool
 	ModuleName     string
-}
-
-type FilterResult any
-
-type Wrapped struct {
-	Wrapped interface{}
-}
-
-var (
-	ErrEmptyWrapped             = errors.New("empty filter result")
-	ErrUnmarshalToTypesNotMatch = errors.New("unmarshal error: input and output types not match")
-)
-
-func (f Wrapped) UnmarhalTo(v any) error {
-	if f.Wrapped == nil {
-		return ErrEmptyWrapped
-	}
-
-	rv := reflect.ValueOf(v)
-	if rv.Kind() != reflect.Pointer || rv.IsNil() {
-		// error replace with "not pointer"
-		return fmt.Errorf("reflect.TypeOf(v): %s", reflect.TypeOf(v))
-	}
-
-	rw := reflect.ValueOf(f.Wrapped)
-	if rw.Kind() != reflect.Pointer || rw.IsNil() {
-		rv.Elem().Set(rw)
-
-		return nil
-	}
-
-	if rw.Type() != rv.Type() {
-		return ErrUnmarshalToTypesNotMatch
-	}
-
-	rv.Elem().Set(rw.Elem())
-
-	return nil
-}
-
-func (f Wrapped) String() string {
-	buf := bytes.NewBuffer([]byte{})
-	_ = json.NewEncoder(buf).Encode(f.Wrapped)
-
-	res := buf.String()
-
-	if strings.HasPrefix(res, "\"") {
-		res = res[1 : len(res)-2]
-	}
-
-	return res
-}
-
-// type Snapshots map[string][]Wrapped
-type Snapshots map[string][]sdkpkg.Snapshot
-
-func (s Snapshots) Get(name string) []sdkpkg.Snapshot {
-	return s[name]
 }
 
 type PatchCollector interface {
@@ -160,8 +96,6 @@ type ScheduleConfig struct {
 	// Crontab is a schedule config in crontab format. (5 or 6 fields)
 	Crontab string
 }
-
-type FilterFunc func(*unstructured.Unstructured) (FilterResult, error)
 
 type KubernetesConfig struct {
 	// Name is a key in snapshots map.
